requirements.txt
Django==2.2.4
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
